📝 备份

+++ +utils/tiff_save_avg.m
@@ -1,5 +1,5 @@
 function imgAvg = tiff_save_avg(filepath,imgStack)
-    imgAvg = mean(imgStack,3);
+    imgAvg = mean(double(imgStack),3);
     imgAvg = im2uint8(mat2gray(imgAvg));
     imwrite(imgAvg,filepath,'Compression','none');
 end
